+++ src/kernel/network/NetworkTransmittable.cpp
@@ -12,13 +12,13 @@ uint8_t NetworkTransmittable::copyTo(Kernel::NetworkByteBlock *output) {
     if (output == nullptr) {
         return 1;
     }
-    size_t lengthInBytes = this->getLengthInBytes();
+    uint16_t lengthInBytes = this->getLengthInBytes();
     if (lengthInBytes == 0 || output->bytesRemaining() < lengthInBytes) {
         return 1;
     }
     return do_copyTo(output);
 }
 
-size_t NetworkTransmittable::getLengthInBytes() {
+uint16_t NetworkTransmittable::getLengthInBytes() {
     return do_getLengthInBytes();
 }
