+++ src/kernel/network/arp/ARPModule.cpp
@@ -65,7 +65,7 @@ namespace Kernel {
     ARPModule::~ARPModule() {
         if (arpTable != nullptr) {
             ARPEntry *toDelete;
-            for (size_t i = 0; i < arpTable->size(); i++) {
+            for (uint32_t i = 0; i < arpTable->size(); i++) {
                 toDelete = arpTable->get(i);
                 arpTable->remove(i);
                 i--;
