+++ src/kernel/network/internet/icmp/ICMP4Echo.cpp
@@ -37,7 +37,7 @@ ICMP4Message::ICMP4MessageType ICMP4Echo::do_getICMP4MessageType() {
     return ICMP4MessageType::ECHO;
 }
 
-size_t ICMP4Echo::do_getLengthInBytes() {
+uint16_t ICMP4Echo::do_getLengthInBytes() {
     return sizeof(header) + sizeof echoMessage;
 }
